refactor: merge activity state into active lease service
