+++ src/main/java/ti4/discord/interactions/listeners/BanListener.java
@@ -8,13 +8,13 @@ import net.dv8tion.jda.api.events.guild.GuildAuditLogEntryCreateEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import ti4.logging.BotLogger;
 import ti4.service.async.BanCleanupService;
-import ti4.spring.service.deploy.InstanceActivityService;
+import ti4.spring.service.deploy.ActiveLeaseService;
 
 class BanListener extends ListenerAdapter {
 
     @Override
     public void onGuildAuditLogEntryCreate(@Nonnull GuildAuditLogEntryCreateEvent event) {
-        if (!InstanceActivityService.shouldHandleCurrentProcessInteraction()) return;
+        if (!ActiveLeaseService.shouldHandleCurrentProcessInteraction()) return;
         try {
             AuditLogEntry log = event.getEntry();
             UserSnowflake target = getTargetUser(log);
