+++ src/main/java/ti4/spring/service/deploy/ActiveLeaseService.java
@@ -12,22 +12,27 @@ import org.springframework.transaction.annotation.Transactional;
 import ti4.logging.BotLogger;
 import ti4.spring.context.SpringContext;
 
+/**
+ * Owns the shared active lease and the local process state that decides whether this instance may serve traffic,
+ * handle Discord interactions, mutate game state, or run scheduled work.
+ */
 @Service
 @RequiredArgsConstructor
 public class ActiveLeaseService {
 
     private final ActiveLeaseRepository activeLeaseRepository;
     private final LeaseProperties leaseProperties;
-    private final InstanceActivityService instanceActivityService;
 
     private final String instanceId = UUID.randomUUID().toString();
+    private final AtomicBoolean active = new AtomicBoolean(false);
+    private final AtomicBoolean draining = new AtomicBoolean(false);
     private final AtomicBoolean ready = new AtomicBoolean(false);
     private final AtomicBoolean leaseParticipationEnabled = new AtomicBoolean(true);
 
     @PostConstruct
     void acquireOnStartup() {
         boolean acquired = tryAcquireLease();
-        instanceActivityService.setActive(acquired);
+        setActive(acquired);
         if (acquired) {
             BotLogger.info("Acquired active lease for instance " + instanceId);
         } else {
@@ -35,6 +40,7 @@ public class ActiveLeaseService {
         }
     }
 
+    /** Attempts to acquire or steal an expired lease for the current process. */
     @Transactional
     public boolean tryAcquireLease() {
         Instant now = Instant.now();
@@ -47,28 +53,31 @@ public class ActiveLeaseService {
         }
 
         activeLeaseRepository.save(buildLease(now));
-        instanceActivityService.setActive(true);
-        instanceActivityService.setDraining(false);
+        setActive(true);
+        setDraining(false);
         return true;
     }
 
+    /** Extends the lease while this process remains active. */
     @Transactional
     public void renewLease() {
-        if (!instanceActivityService.isActive()) {
+        if (!isActive()) {
             return;
         }
         activeLeaseRepository.save(buildLease(Instant.now()));
     }
 
+    /** Releases the lease if it is currently owned by this process. */
     @Transactional
     public void releaseLease() {
         Optional<ActiveLeaseEntity> existing = activeLeaseRepository.findById(leaseProperties.getLeaseKey());
         if (existing.isPresent() && instanceId.equals(existing.get().getInstanceId())) {
             activeLeaseRepository.delete(existing.get());
         }
-        instanceActivityService.setActive(false);
+        setActive(false);
     }
 
+    /** Returns whether the shared lease row is still owned by this process. */
     public boolean stillOwnsLease() {
         return activeLeaseRepository
                 .findById(leaseProperties.getLeaseKey())
@@ -76,52 +85,86 @@ public class ActiveLeaseService {
                 .orElse(false);
     }
 
+    /** Returns whether this process may perform persistent game mutations right now. */
     public boolean mayMutate() {
-        return instanceActivityService.isActive() && stillOwnsLease();
+        return isActive() && stillOwnsLease();
     }
 
-    public boolean shouldServeTraffic() {
-        return isReady() && mayMutate() && !instanceActivityService.isDraining();
+    /** Returns whether this process is the currently active instance. */
+    public boolean isActive() {
+        return active.get();
     }
 
+    /** Marks this process active or inactive, clearing drain state when deactivating. */
+    public void setActive(boolean active) {
+        this.active.set(active);
+        if (!active) {
+            draining.set(false);
+        }
+    }
+
+    /** Returns whether this process is draining and refusing new work. */
+    public boolean isDraining() {
+        return draining.get();
+    }
+
+    /** Marks this process as draining or not draining. */
+    public void setDraining(boolean draining) {
+        this.draining.set(draining);
+    }
+
+    /** Returns whether this process should handle Discord interaction ingress. */
+    public boolean shouldHandleDiscordInteraction() {
+        return isActive() && !isDraining();
+    }
+
+    /** Returns whether startup has completed for this process. */
     public boolean isReady() {
         return ready.get();
     }
 
+    /** Marks this process ready or not ready for normal traffic. */
     public void setReady(boolean ready) {
         this.ready.set(ready);
     }
 
+    /** Returns whether this process should currently serve normal HTTP traffic. */
+    public boolean shouldServeTraffic() {
+        return isReady() && mayMutate() && !isDraining();
+    }
+
+    /** Initiates drain and closes the Spring context on a background thread. */
     public boolean requestDrain() {
-        if (!instanceActivityService.isActive() || instanceActivityService.isDraining()) {
+        if (!isActive() || isDraining()) {
             return false;
         }
 
         leaseParticipationEnabled.set(false);
-        instanceActivityService.setDraining(true);
+        setDraining(true);
         setReady(false);
         BotLogger.info("Drain requested for active instance " + instanceId);
         Thread.ofPlatform().name("bot-drain-shutdown").start(SpringContext::closeApplicationContext);
         return true;
     }
 
+    /** Returns the randomly generated identifier used for this process in the lease row. */
     public String currentInstanceId() {
         return instanceId;
     }
 
     @Scheduled(fixedDelayString = "#{@leaseProperties.heartbeatIntervalMillis}")
     void maintainLease() {
-        if (instanceActivityService.isDraining()) {
+        if (isDraining()) {
             return;
         }
 
-        if (instanceActivityService.isActive()) {
+        if (isActive()) {
             if (stillOwnsLease()) {
                 renewLease();
             } else {
                 BotLogger.warning("Active instance lost lease ownership: " + instanceId);
                 setReady(false);
-                instanceActivityService.setActive(false);
+                setActive(false);
             }
             return;
         }
@@ -149,6 +192,7 @@ public class ActiveLeaseService {
         return lease.getLeaseExpiresAt() == null || lease.getLeaseExpiresAt().isBefore(now);
     }
 
+    /** Returns current-process readiness, defaulting to false before Spring has initialized. */
     public static boolean isCurrentProcessReady() {
         try {
             return SpringContext.getBean(ActiveLeaseService.class).isReady();
@@ -157,6 +201,7 @@ public class ActiveLeaseService {
         }
     }
 
+    /** Updates current-process readiness if Spring has initialized. */
     public static void setCurrentProcessReady(boolean ready) {
         try {
             SpringContext.getBean(ActiveLeaseService.class).setReady(ready);
@@ -165,6 +210,16 @@ public class ActiveLeaseService {
         }
     }
 
+    /** Returns whether the current process should handle a Discord interaction. */
+    public static boolean shouldHandleCurrentProcessInteraction() {
+        try {
+            return SpringContext.getBean(ActiveLeaseService.class).shouldHandleDiscordInteraction();
+        } catch (IllegalStateException e) {
+            return false;
+        }
+    }
+
+    /** Returns whether the current process should serve normal HTTP traffic. */
     public static boolean shouldCurrentProcessServeTraffic() {
         try {
             return SpringContext.getBean(ActiveLeaseService.class).shouldServeTraffic();
@@ -173,11 +228,11 @@ public class ActiveLeaseService {
         }
     }
 
+    /** Returns whether the current process should run active-instance scheduled work. */
     public static boolean shouldCurrentProcessRunScheduledWork() {
         try {
             ActiveLeaseService activeLeaseService = SpringContext.getBean(ActiveLeaseService.class);
-            return activeLeaseService.mayMutate()
-                    && !SpringContext.getBean(InstanceActivityService.class).isDraining();
+            return activeLeaseService.mayMutate() && !activeLeaseService.isDraining();
         } catch (IllegalStateException e) {
             return false;
         }
